+++ supabase/migrations/0120_fix_stng_resolver_required_validation.sql
@@ -345,7 +345,7 @@ COMMIT;
 
 -- Example test query:
 -- SELECT * FROM fn_stng_resolve_setting_value(
---   '11111111-1111-1111-1111-111111111111'::UUID,
+--   '11111111-1111-4111-8111-111111111111'::UUID,
 --   'TENANT_CURRENCY',
 --   NULL,
 --   NULL